Kotlin

+++ 0x0B-malloc_free/3-alloc_grid.c
@@ -15,14 +15,14 @@ int **alloc_grid(int width, int height)
 	if (width <= 0 || height <= 0)
 		return (NULL);
 
-	array = malloc(sizeof(int *) * width);
+	array = malloc(sizeof(long *) * width);
 
 	if (array == NULL)
 		return (NULL);
 
 	for (i = 0; i < height; i++)
 	{
-		array[i] = malloc(sizeof(int) * height);
+		array[i] = malloc(sizeof(long) * height);
 		if (array[i] == NULL)
 			return (NULL);
 	}
